Misc: run forge fmt

+++ src/interfaces/IWETH.sol
@@ -7,5 +7,5 @@ import "./IERC20.sol";
 
 interface IWETH is IERC20 {
     function deposit() external payable;
-    function withdraw(uint amount) external;
+    function withdraw(uint256 amount) external;
 }
